--- components/lie_amber/tests/__init__.py
@@ -1,21 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""
-Python unit test function for lie_atb module, run as:
-::
-    test = module_test_suite()
-    runner = unittest.TextTestRunner(verbosity=2)
-    runner.run(test)
-"""
-
-import os
-import unittest2
-
-def module_test_suite():
-    """
-    Run lie_atb module unit tests
-    """
-    
-    loader = unittest2.TestLoader()
-    suite = loader.discover(os.path.dirname(__file__), pattern='module_*.py')
-    return suite
